+++ src/containerapp/HISTORY.rst
@@ -7,6 +7,7 @@ Upcoming
 * 'az containerapp env update': Fix bugs in update environment.
 * Fix YAML create with user-assigned identity
 * Fix polling logic for long running operations.
+* Add `az containerapp ingress update` Command to Update Container App Ingress
 
 0.3.24
 ++++++